// old/src/types/course-d-t.ts
export type ICourseDT = {
  id: number;
  title: string;
  author_img?: string;
  author_name: string;
  thumbnail: string;
  category: string;
  price: number;
  old_price?: number;
  lectures_count: number;
  duration: number;
  level: number;
  avg_rating: number;
  discount: number;
  lessons: number;
  students: number;
  total_rating: number;
  language: string;
  progress?: number;
  course_: string;
  created_at?: string;
  instructor_name?: string;
  rating?: string;
  slug: string;
  description?: string;
};